NetworkTroubleShooting-Agent/requirements.txt
# ========================
# BlueCom Network Troubleshooter - Python Dependencies
# ========================
# Install with: pip install -r requirements.txt

# ========================
# Core Dependencies
# ========================
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2

# ========================
# Vector Database & Search
# ========================
qdrant-client==1.7.0

# ========================
# Machine Learning & NLP
# ========================
# Embeddings and transformers
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2

# Tokenizers
tokenizers==0.15.0

# Additional ML utilities
scikit-learn==1.3.2
scipy==1.11.4

# ========================
# LangChain & LangGraph
# ========================
langgraph==0.0.20
langchain==0.1.0
langchain-core==0.1.10
langchain-community==0.0.13

# ========================
# AWS Services
# ========================
boto3==1.34.10
botocore==1.34.10
s3transfer==0.10.0

# ========================
# Web Interface
# ========================
streamlit==1.29.0
streamlit-extras==0.3.6

# Streamlit dependencies
altair==5.2.0
click==8.1.7
packaging==23.2
pillow==10.1.0
protobuf==4.25.1
pyarrow==14.0.1
pydeck==0.8.1b0
requests==2.31.0
toml==0.10.2
tornado==6.4
tzlocal==5.2
watchdog==3.0.0

# ========================
# HTTP & API
# ========================
urllib3==2.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6

# ========================
# Data Processing
# ========================
# Date and time
python-dateutil==2.8.2
pytz==2023.3.post1

# Text processing
regex==2023.12.25

# File handling
openpyxl==3.1.2

# ========================
# Utilities
# ========================
# Type checking
typing-extensions==4.9.0

# Progress bars
tqdm==4.66.1

# YAML support
PyYAML==6.0.1

# JSON handling
jsonschema==4.20.0

# ========================
# Hugging Face
# ========================
huggingface-hub==0.19.4
safetensors==0.4.1

# ========================
# Optional: Development Tools
# ========================
# Uncomment if needed for development

# Testing
# pytest==7.4.3
# pytest-cov==4.1.0

# Code quality
# black==23.12.1
# flake8==7.0.0
# mypy==1.7.1

# Jupyter (for notebooks)
# jupyter==1.0.0
# ipykernel==6.27.1

# ========================
# Platform-Specific Notes
# ========================
# Windows: Install Visual C++ Build Tools if torch installation fails
# Mac M1/M2: Use torch with MPS support
# Linux: Standard installation should work

# ========================
# Version Compatibility
# ========================
# Python: 3.9, 3.10, 3.11
# Tested on: Python 3.9.18, 3.10.13, 3.11.7

# ========================
# Installation Notes
# ========================
# 1. Create virtual environment:
#    python -m venv venv
#
# 2. Activate virtual environment:
#    Windows: venv\Scripts\activate
#    Linux/Mac: source venv/bin/activate
#
# 3. Install dependencies:
#    pip install -r requirements.txt
#
# 4. First run will download ~2GB of models:
#    - SentenceTransformer: all-MiniLM-L6-v2 (~80MB)
#    - Google Gemma 2B-Instruct (~4.5GB)
#
# 5. Verify installation:
#    python -c "import torch; import transformers; import qdrant_client; print('OK')"

# ========================
# Known Issues
# ========================
# Issue: torch installation fails on Windows
# Solution: Install Visual C++ Build Tools
# https://visualstudio.microsoft.com/visual-cpp-build-tools/
#
# Issue: Out of memory loading Gemma 2B
# Solution: Ensure 8GB+ RAM available or use smaller model
#
# Issue: qdrant-client connection timeout
# Solution: Check firewall settings and QDRANT_URL in .env